requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
dataclasses-json==0.5.6
dulwich==0.20.26
feedparser==6.0.8
grimoirelab-toolkit==0.2.0
idna==3.3
imagesize==1.2.0
markdownify==0.9.4
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
perceval==0.17.13
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
python-dateutil==2.8.2
python-magic==0.4.24
requests==2.26.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3
text-unidecode==1.3
typing-extensions==4.0.0
typing-inspect==0.7.1
urllib3==1.26.7
